Basic cutter implementation New document/view type for driving cutter - CNC related classes controlled by that with plotting managed by existing doc view.  Various changes to support that including GCode handling and DXF import.

+++ Kernel/Wing.cpp
@@ -158,7 +158,7 @@ void CWing::serializeTo(CObjectSerializer& os)
 	
 	flags.serializeTo(os);
 
-	os.startCollection("spars",spars.size());
+	os.startCollection("spars",(int)spars.size());
 	for(SPARS::iterator iter = spars.begin();
 	iter != spars.end();
 	++iter)
